+++ poem-openapi/src/error.rs
@@ -20,15 +20,15 @@ impl ResponseError for ParseParamError {
     }
 }
 
-/// Parse JSON error.
+/// Parse request payload error.
 #[derive(Debug, Error)]
-#[error("parse JSON error: {reason}")]
-pub struct ParseJsonError {
+#[error("parse request payload error: {reason}")]
+pub struct ParseRequestPayloadError {
     /// The reason for the error.
     pub reason: String,
 }
 
-impl ResponseError for ParseJsonError {
+impl ResponseError for ParseRequestPayloadError {
     fn status(&self) -> StatusCode {
         StatusCode::BAD_REQUEST
     }