+++ poem-openapi/src/payload/mod.rs
@@ -1,6 +1,7 @@
 //! Commonly used payload types.
 
 mod attachment;
+mod base64_payload;
 mod binary;
 mod event_stream;
 mod html;
@@ -11,8 +12,8 @@ mod response;
 use poem::{Request, RequestBody, Result};
 
 pub use self::{
-    attachment::Attachment, binary::Binary, event_stream::EventStream, html::Html, json::Json,
-    plain_text::PlainText, response::Response,
+    attachment::Attachment, base64_payload::Base64, binary::Binary, event_stream::EventStream,
+    html::Html, json::Json, plain_text::PlainText, response::Response,
 };
 use crate::registry::{MetaSchemaRef, Registry};